+++ src/store/store.js
@@ -450,14 +450,17 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
       if (!_request) {
         return callback(store);
       }
-      var _cacheValues = cache.get(store.name, md5(_request.request));
+
+      var _cacheValues = cache.get(_options.store.name, md5(_request.request));
 
       if (_cacheValues) {
         if (typeof _cacheValues === 'object') {
           afterAction(_options.store, 'get', _transformGetCache(_options.collection, _cacheValues), null, _transactionId);
-          return callback(store);
         }
-        afterAction(_options.store, 'get', [], null, _transactionId);
+        else {
+          afterAction(_options.store, 'get', [], null, _transactionId);
+        }
+        hook.pushToHandlers(_options.store, 'filterUpdated', false, null, _transactionId);
         return callback(store);
       }
 
@@ -520,7 +523,7 @@ function _get (store, primaryKeyValue, retryOptions, callback) {
       afterAction(_options.store, 'get', data, null, _transactionId);
       _propagate(_options.store, data, utils.OPERATIONS.INSERT, _transactionId);
       if (_options.store.isFilter) {
-        hook.pushToHandlers(_options.store, 'filterUpdated');
+        hook.pushToHandlers(_options.store, 'filterUpdated', false, null, _transactionId);
       }
       storeUtils.saveState(_options.store, _options.collection, _options.cache);
     });